fix: ANT-5229 - fix liquibase syntax and turn reservoir.cap to decimal

+++ src/main/java/com/rte_france/antares/datamanager_back/dto/HydroPropertiesGenerationDTO.java
@@ -30,7 +30,7 @@ public class HydroPropertiesGenerationDTO {
     private Boolean reservoirManagement;
 
     @JsonProperty("reservoir_capacity")
-    private Integer reservoirCapacity;
+    private BigDecimal reservoirCapacity;
 
     @JsonProperty("pumping_efficiency")
     private BigDecimal pumpingEfficiency;
